fix(core): make sure ignore strategies receive directories with trailing slash
ignore strategies like git, expect directories to be passed in with a
trailing slash in order to properly match.

consider a source tree:

├── hello
│   └── index.html
├── index.html

now we want to exclude everything except all html files
(including all subdirectories).

this can be done with a gitignore of:
*
!*/
!*.html

in order for the git ignore strategy to properly match this,
it needs to know that hello is a directory.

this patch just does that by appending a trailing slash for the
ignore strategies.